GranularLammpsDump Documentation

GranularLammpsDump Documentation

Shape/Parse/Export Dump Files

Index

Shape/Parse/Export Dump Files

GranularLammpsDump.getNatoms — Method

getNatoms("inputfile")

Parses the total number of atoms into an integer.

Looks at the 4th line of the dump file and parses the number of atoms from the readline string into an integer.

GranularLammpsDump.readdump — Method

readdump("inputfile")

Converts the contents of the dumpfile to a massive dictionary.

Reads the dumpfile into a matrix, shapes the matrix, and removes text lines. The number of atoms is extracted and used internally but not returned. The final dictionary form has timesteps as keys, and matrices sorted by particle ID number as values. In principle, this works for any dumpfile with at least 6 columns of output, but has been tested with 6 output columns and 9 output columns.

> Each timestep looks something like this: 
> | 1  |  2    | 3 | 4 | 5 | 6  | 7  | 8  | 9  | 10 | 11 |
> | ID | GROUP | x | y | z | vx | vy | vz | ux | uy | uz |

GranularLammpsDump.parsestep — Method

parsestep(d,ts)

Stores critical info from a dump step in a mutable struct.

Iterate this in a loop to step through a simulation.

GranularLammpsDump.dump2mat — Function

dump2mat(dict,exportflag=1)

Converts the readdump dictionary to one giant matrix. Optionally export as .csv file.

Export to .csv is on by default. This will help you collaborate with Matlab users ;) Set exportflag to 0 to turn off.

Visualize

GranularLammpsDump.setdefaults — Method

setdefaults()

Create the first settings file for the project.

Example

julia> cd("/home/JuliaUser/Project")
julia> setdefaults()  
Default settings.conf created.  
Edit this using menu() or your favorite text editor.  

GranularLammpsDump.menu — Method

menu()

Brings up the settings menu in system program nano.

Sneaky dependency: requires nano to be installed prior to use.

GranularLammpsDump.settingsloader — Method

settingsloader()

Loads the settings.conf file into a settings dictionary.

Used inside other functions, but not useful as a standalone function.

GranularLammpsDump.makemovie_allgrains — Function

makemovie_allgrains(dump, skips, dumpfile="allgrains")

Makes an mp4 movie with 5 groups color coded by initial x position.

The "skips" variable takes every Nth timestep from the dump dictionary. E.g., use 1 for every timestep or 10 to take every 10th step.

Output is an .mp4 video in the project directory. The optional dumpfile argument allows matching the output.mp4 to the input file name.

GranularLammpsDump.makemovie_xslice — Function

makemovie_xslice(dump, skips, dumpfile="allgrains")

Makes an mp4 movie from one x slice, top to bottom, and color codes the particles based on "granular temperature" in the current step.

The "skips" variable takes every Nth timestep from the dump dictionary. E.g., use 1 for every timestep or 10 to take every 10th step.

Output is an .mp4 video in the project directory. The optional dumpfile argument allows matching the output.mp4 to the input file name.

Index

@index