Visualizing Dump Files

Initial Set Up

GranularLammpsDump.setdefaults — Method

setdefaults()

Create the first settings file for the project.

Example

julia> cd("/home/JuliaUser/Project")
julia> setdefaults()  
Default settings.conf created.  
Edit this using menu() or your favorite text editor.  

GranularLammpsDump.menu — Method

menu()

Brings up the settings menu in system program nano.

Sneaky dependency: requires nano to be installed prior to use.

GranularLammpsDump.settingsloader — Method

settingsloader()

Loads the settings.conf file into a settings dictionary.

Used inside other functions, but not useful as a standalone function.

Color coded by initial x position

GranularLammpsDump.makemovie_allgrains — Function

makemovie_allgrains(dump,boxes,skips,moviename="granularmovie")

Makes an .mp4 movie with all grains shown, color coded by initial x position.

Color coded by granular temperature

GranularLammpsDump.makemovie — Function

makemovie(dump,boxes,Tref,skips,moviename="granularmovie")

Makes an .mp4 movie with all grains shown depicting temperature.

Grains outside the slice of interest are shown as transparent with black borders.

ARGUMENTS

Tref is a string, like "0:1/4:6" that sets the powers in 10^Tref